[CAP-49] Add logging to driver, flight path, and NOTAM fetching
Uses python's logging module. Logs write to the terminal as well as
output.log. This log is added to the gitignore so it should not be
commited to the repo.

By default, logging is at the DEBUG level.

For fetching NOTAMs, print the total NOTAMs fetched so far, and record the
total time taken to fetch all notams.